Revert "Shortcuts (#21083)"
This reverts commit 6e7f3e6f5a11341edd5c2519cbe1c7bc178cb8e9.
