+++ packages/flutter/lib/src/services/text_input.dart
@@ -534,23 +534,6 @@ class TextEditingValue {
   );
 }
 
-/// An interface for manipulating the selection, to be used by the implementor
-/// of the toolbar widget.
-abstract class TextSelectionDelegate {
-  /// Gets the current text input.
-  TextEditingValue get textEditingValue;
-
-  /// Sets the current text input (replaces the whole line).
-  set textEditingValue(TextEditingValue value);
-
-  /// Hides the text selection toolbar.
-  void hideToolbar();
-
-  /// Brings the provided [TextPosition] into the visible area of the text
-  /// input.
-  void bringIntoView(TextPosition position);
-}
-
 /// An interface to receive information from [TextInput].
 ///
 /// See also:
