+++ packages/flutter/lib/src/widgets/editable_text.dart
@@ -904,7 +904,6 @@ class EditableTextState extends State<EditableText> with AutomaticKeepAliveClien
               rendererIgnoresPointer: widget.rendererIgnoresPointer,
               cursorWidth: widget.cursorWidth,
               cursorRadius: widget.cursorRadius,
-              textSelectionDelegate: this,
             ),
           ),
         );
@@ -968,7 +967,6 @@ class _Editable extends LeafRenderObjectWidget {
     this.rendererIgnoresPointer = false,
     this.cursorWidth,
     this.cursorRadius,
-    this.textSelectionDelegate,
   }) : assert(textDirection != null),
        assert(rendererIgnoresPointer != null),
        super(key: key);
@@ -992,7 +990,6 @@ class _Editable extends LeafRenderObjectWidget {
   final bool rendererIgnoresPointer;
   final double cursorWidth;
   final Radius cursorRadius;
-  final TextSelectionDelegate textSelectionDelegate;
 
   @override
   RenderEditable createRenderObject(BuildContext context) {
@@ -1015,7 +1012,6 @@ class _Editable extends LeafRenderObjectWidget {
       obscureText: obscureText,
       cursorWidth: cursorWidth,
       cursorRadius: cursorRadius,
-      textSelectionDelegate: textSelectionDelegate,
     );
   }
 
@@ -1039,7 +1035,6 @@ class _Editable extends LeafRenderObjectWidget {
       ..ignorePointer = rendererIgnoresPointer
       ..obscureText = obscureText
       ..cursorWidth = cursorWidth
-      ..cursorRadius = cursorRadius
-      ..textSelectionDelegate = textSelectionDelegate;
+      ..cursorRadius = cursorRadius;
   }
 }
