+++ packages/flutter/lib/src/widgets/text_selection.dart
@@ -16,8 +16,6 @@ import 'gesture_detector.dart';
 import 'overlay.dart';
 import 'transitions.dart';
 
-export 'package:flutter/services.dart' show TextSelectionDelegate;
-
 /// Which type of selection handle to be displayed.
 ///
 /// With mixed-direction text, both handles may be the same type. Examples:
@@ -61,6 +59,23 @@ enum _TextSelectionHandlePosition { start, end }
 ///
 /// Used by [TextSelectionOverlay.onSelectionOverlayChanged].
 typedef void TextSelectionOverlayChanged(TextEditingValue value, Rect caretRect);
+
+/// An interface for manipulating the selection, to be used by the implementor
+/// of the toolbar widget.
+abstract class TextSelectionDelegate {
+  /// Gets the current text input.
+  TextEditingValue get textEditingValue;
+
+  /// Sets the current text input (replaces the whole line).
+  set textEditingValue(TextEditingValue value);
+
+  /// Hides the text selection toolbar.
+  void hideToolbar();
+
+  /// Brings the provided [TextPosition] into the visible area of the text
+  /// input.
+  void bringIntoView(TextPosition position);
+}
 
 /// An interface for building the selection UI, to be provided by the
 /// implementor of the toolbar widget.
